+++ src/irvine/oeis/a344/A344523.java
@@ -14,6 +14,6 @@ public class A344523 extends LambdaSequence {
 
   /** Construct the sequence. */
   public A344523() {
-    super(1, n -> Integers.SINGLETON.sum(1, n, k -> Functions.PHI.z((long) k).multiply(Z.valueOf(n / k).pow(4))));
+    super(1, n -> Integers.SINGLETON.sum(1, n, k -> Functions.PHI.z(k).multiply(Z.valueOf(n / k).pow(4))));
   }
 }
