+++ src/irvine/oeis/a344/A344525.java
@@ -14,6 +14,6 @@ public class A344525 extends LambdaSequence {
 
   /** Construct the sequence. */
   public A344525() {
-    super(1, n -> Integers.SINGLETON.sum(1, n, k -> Functions.PHI.z((long) k).multiply(Z.valueOf(n / k).pow(n))));
+    super(1, n -> Integers.SINGLETON.sum(1, n, k -> Functions.PHI.z(k).multiply(Z.valueOf(n / k).pow(n))));
   }
 }
